fix(vdsnapshot): resolve fsfreeze race condition (#1668)
Sometimes, when the virtual disk snapshot controller creates a batch of snapshots with required consistency, a race condition in the freeze or unfreeze "filesystem" operation can cause an inconsistent snapshot. To prevent this, the KubeVirt virtual machine instance is now annotated with a "freeze" or "unfreeze" request, and the service checks that the type of this request matches the "fsFreezeStatus". The "fsFreezeStatus" is considered trusted only if it matches the request type.

+++ test/e2e/legacy/vd_snapshots.go
@@ -30,6 +30,7 @@ import (
 
 	"github.com/deckhouse/virtualization/api/core/v1alpha2"
 	"github.com/deckhouse/virtualization/api/core/v1alpha2/vmcondition"
+	"github.com/deckhouse/virtualization/test/e2e/internal/config"
 	kc "github.com/deckhouse/virtualization/test/e2e/internal/kubectl"
 	"github.com/deckhouse/virtualization/test/e2e/internal/util"
 )
@@ -360,19 +361,21 @@ var _ = Describe("VirtualDiskSnapshots", Ordered, func() {
 
 	Context("When test is completed", func() {
 		It("deletes test case resources", func() {
-			DeleteTestCaseResources(ns, ResourcesToDelete{
-				KustomizationDir: conf.TestData.VdSnapshots,
-				AdditionalResources: []AdditionalResource{
-					{
-						Resource: kc.ResourceVDSnapshot,
-						Labels:   hasNoConsumerLabel,
+			if config.IsCleanUpNeeded() {
+				DeleteTestCaseResources(ns, ResourcesToDelete{
+					KustomizationDir: conf.TestData.VdSnapshots,
+					AdditionalResources: []AdditionalResource{
+						{
+							Resource: kc.ResourceVDSnapshot,
+							Labels:   hasNoConsumerLabel,
+						},
+						{
+							Resource: kc.ResourceVDSnapshot,
+							Labels:   attachedVirtualDiskLabel,
+						},
 					},
-					{
-						Resource: kc.ResourceVDSnapshot,
-						Labels:   attachedVirtualDiskLabel,
-					},
-				},
-			})
+				})
+			}
 		})
 	})
 })
